\name{AbrevSp}
\alias{AbrevSp}
\title{AbrevSp}
\usage{
AbrevSp(sp,size)
}
\description{
Create abrevation for species or genra.
}

\arguments{
\item{sp}{Table colonm where there are the name of species or genra}

\item{size}{Size of abreviation, number of letter in abreviation}



\author{
Augustin Soulard
}
