\name{dissimilarity}
\alias{dissimilarity}
\title{dissimilarity}
\usage{
dissimilarity(site1,site2,method = "Sorenson")
}
\description{
Calculate dissimilarity index between two site.
}

\arguments{
\item{site1}{Row for first site}

\item{site2}{Row for second site}

\item{method}{Dissimilarity method returned : "Sorenson", "Simpson" or "Nestedness"}

\author{
Augustin Soulard
}

\references{
Baselga, A. (2010). Partitioning the turnover and nestedness components of beta diversity. Global ecology and biogeography, 19(1), 134-143.
}
