\name{sptogenra}
\alias{sptogenra}
\title{sptogenra}
\usage{
sptogenra(sp,sep="_")
}
\description{
Converte row names of species in genra and sum species of same genra.
}

\arguments{
\item{sp}{Table with row names are species}

\item{sep}{Separator between species and genra in species names}



\author{
Augustin Soulard
}
