\name{turn_over}
\alias{turn_over}
\title{turn_over}
\usage{
turn_over(site1,site2,title)
}
\description{
Genera table witch differentiate number of species on each site and species common for 2 sites
}

\arguments{
\item{site1}{Row for first site}

\item{site2}{Row for second site}

\item{title}{titles of tab colomns}

\author{
Augustin Soulard
}
